+++ pagegraph/commands/js_calls.py
@@ -11,6 +11,7 @@ if TYPE_CHECKING:
     from pathlib import Path
     from typing import Optional, Union
 
+    from pagegraph.node.dom_root import DOMRootNode
     from pagegraph.serialize import ScriptReport, BasicReport, JSCallResultReport
     from pagegraph.types import PageGraphId
 
@@ -36,7 +37,7 @@ class Command(pagegraph.commands.Base):
         self.pg_id = pg_id
         super().__init__(input_path, debug)
 
-    def validate(self) -> bool:
+    def validate(self) -> None:
         if self.frame_nid:
             pagegraph.commands.validate_node_id(self.frame_nid)
         if self.pg_id:
@@ -46,6 +47,27 @@ class Command(pagegraph.commands.Base):
     def execute(self) -> pagegraph.commands.Result:
         pg = pagegraph.graph.from_path(self.input_path, self.debug)
         reports: list[Result] = []
+
+        domroot_node: Optional[DOMRootNode] = None
+        if self.frame_nid:
+            domroot_node = pg.node(self.frame_nid).as_domroot_node()
+            if not domroot_node:
+                raise ValueError("The PageGraph id provided is not " +
+                    f"a DOMRootNode, nid={self.frame_id}")
+        assert domroot_node
+
+        can_do_fast_path = (domroot_node is not None and
+            not self.cross_frame and not self.pg_id and not self.method)
+        if can_do_fast_path:
+            for call_edge in pg.js_call_edges():
+                if call_edge.frame_id() != domroot_node.frame_id():
+                    continue
+                js_result = call_edge.call_result()
+                result_report = js_result.to_report()
+                script_node = js_result.call.incoming_node()
+                script_report = script_node.to_report()
+                reports.append(Result(script_report, result_report))
+            return pagegraph.commands.Result(pg, reports)
 
         js_structure_nodes = pg.js_structure_nodes()
         for js_node in js_structure_nodes: